+++ package.json
@@ -26,12 +26,6 @@
     "signin": "office-addin-dev-settings m365-account login",
     "signout": "office-addin-dev-settings m365-account logout",
     "start": "office-addin-debugging start manifest.xml",
-    "start:desktop": "office-addin-debugging start manifest.xml desktop",
-    "start:desktop:word": "office-addin-debugging start manifest.xml desktop --app word",
-    "start:desktop:excel": "office-addin-debugging start manifest.xml desktop --app excel",
-    "start:desktop:powerpoint": "office-addin-debugging start manifest.xml desktop --app powerpoint",
-    "start:desktop:outlook": "office-addin-debugging start manifest.xml desktop --app outlook",
-    "start:web": "office-addin-debugging start manifest.xml web",
     "stop": "office-addin-debugging stop manifest.xml",
     "test": "npm run test:unit && npm run test:e2e",
     "test:e2e": "mocha -r ts-node/register test/end-to-end/*.ts",
