+++ src/commands/commands.outlook.ts
@@ -1,6 +1,31 @@
-import { setNotificationInOutlook } from "./outlook";
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+ * See LICENSE in the project root for license information.
+ */
 
 /* global Office */
+
+/**
+ * Show an outlook notification when the add-in command is executed.
+ * @param event
+ */
+export function setNotificationInOutlook(event: Office.AddinCommands.Event) {
+  const message: Office.NotificationMessageDetails = {
+    type: Office.MailboxEnums.ItemNotificationMessageType.InformationalMessage,
+    message: "Performed action.",
+    icon: "Icon.80x80",
+    persistent: true,
+  };
+
+  // Show a notification message.
+  Office.context.mailbox.item.notificationMessages.replaceAsync(
+    "ActionPerformanceNotification",
+    message
+  );
+
+  // Be sure to indicate when the add-in command function is complete.
+  event.completed();
+}
 
 // Register the add-in commands with the Office host application.
 Office.onReady(async () => {
