+++ src/commands/commands.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
+ * See LICENSE in the project root for license information.
+ */
+
+import "./commands.word";
+import "./commands.excel";
+import "./commands.powerpoint";
+import "./commands.outlook";
