--- src/commands/excel.ts
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
- * See LICENSE in the project root for license information.
- */
-
-/* global Office Excel console */
-
-/**
- * Set range color to selected range in excel when the add-in command is executed.
- * @param event
- */
-export async function setRangeColorInExcel(event: Office.AddinCommands.Event) {
-  try {
-    await Excel.run(async (context) => {
-      const range = context.workbook.getSelectedRange();
-      range.format.fill.color = "yellow";
-      await context.sync();
-    });
-  } catch (error) {
-    // Note: In a production add-in, notify the user through your add-in's UI.
-    console.error(error);
-  }
-
-  // Be sure to indicate when the add-in command function is complete
-  event.completed();
-}
